+++ packages/core/webpack/src/plugins/dtsBundleGeneratorPlugin.ts
@@ -3,14 +3,11 @@
  */
 
 import WebpackShellPluginNextPlugin from "webpack-shell-plugin-next";
-import { rootDir } from "../utils";
 
-const entry = [
-  `node ${rootDir}/node_modules/.bin/dts-bundle-generator -o ./dist/index.d.ts ./src/index.ts`,
-];
+const entry = ["yarn g:dts -o ./dist/index.d.ts ./src/index.ts"];
 
 export const dtsBundleGeneratorPlugin = (
-  configuration: { entries: string[] } = { entries: entry }
+  configuration: { entries: string[] } = { entries: entry },
 ) => {
   const config = {
     onBuildEnd: {