+++ packages/implementations/react/src/atomic/Input/src/index.ts
@@ -2,4 +2,5 @@ import { Input } from "./Input";
 
 export { Input } from "./Input";
 export type { InputProps } from "./Input";
+export type { InputProperties } from "./input.types";
 export default Input;