+++ packages/implementations/react/src/composite/Datepicker/src/components/Input/input.types.ts
@@ -1,0 +1,6 @@
+import { InputHTMLAttributes } from "react";
+
+export interface InputProperties {}
+
+export type InputProps = InputProperties &
+  Omit<InputHTMLAttributes<HTMLInputElement>, "size">;