+++ packages/implementations/react/src/composite/Datepicker/src/index.ts
@@ -1,0 +1,5 @@
+import { Datepicker } from "./Datepicker";
+
+export { Datepicker } from "./Datepicker";
+export type { DatepickerProps } from "./datepicker.types";
+export default Datepicker;