+++ packages/styles/src/packages/composite/datepicker/index.ts
@@ -1,0 +1,5 @@
+import { datepickerStyles } from "./rarui-datepicker.css";
+
+export const datepicker = {
+  classnames: datepickerStyles,
+};